# gateway-service/docs/swagger.yaml
basePath: /api/v1
host: localhost:8080
info:
  contact: {}
  description: API Gateway for Delivery Management System
  title: Delivery Management API Gateway
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Get detailed information of the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get user information
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return token
      parameters:
      - description: Login information
        in: body
        name: user
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Login user
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: Register a new user in the system
      parameters:
      - description: User Registration Info
        in: body
        name: user
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
      summary: Create a new user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Enter your bearer token in the format **Bearer + token**
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"